refactor: centralize error policy adapters
